% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeGibbs.R
\name{check_IDs}
\alias{check_IDs}
\title{Check whether 2 records which are estimated to be linked have the same IDs}
\usage{
check_IDs(recpair, identity_vector)
}
\arguments{
\item{recpair}{A record pair}

\item{identity_vector}{A vector of the unique ids}
}
\value{
Whether or not two records which are estimated to be linked have the same unique ids
}
\description{
Check whether 2 records which are estimated to be linked have the same IDs
}

