% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/decorate.R
\name{decorate}
\alias{decorate}
\title{Decorate the slow function to enable caching}
\usage{
decorate(fun, salt, type = "type", id_col = "id")
}
\description{
Decorate the slow function to enable caching
}

