% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{identity.RLdata500}
\alias{identity.RLdata500}
\title{identity.RLdata500}
\format{
An object of class \code{numeric} of length 500.
}
\usage{
identity.RLdata500
}
\description{
Unique identifier for RLdata500 the corresponds to the record number
format A vector that contains the code{id}
}
\keyword{datasets}
