% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeGibbs.R
\name{links.compare}
\alias{links.compare}
\title{This function takes a set of pairwise links and identifies correct, incorrect,
and missing links
(correct = estimated and true, incorrect = estimated but not true,
missing = true but not estimated)}
\usage{
links.compare(est.links.pair, true.links.pair, counts.only = TRUE)
}
\arguments{
\item{est.links.pair}{The number of estimated links}

\item{true.links.pair}{The number of true links}

\item{counts.only}{State whether or not counts only is true or false}
}
\value{
Gives a vector of the estimated and true links, estimated but not true links, and the true but not estimated links
}
\description{
This function takes a set of pairwise links and identifies correct, incorrect,
and missing links
(correct = estimated and true, incorrect = estimated but not true,
missing = true but not estimated)
}
