% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeGibbs.R
\name{mms}
\alias{mms}
\title{Function to compute a record's Maximal Matching Set (MMS) based on a single linkage structure}
\usage{
mms(lambda, record)
}
\arguments{
\item{lambda}{The linkage structure}

\item{record}{A vector of records}
}
\value{
Computes a records MMS
}
\description{
Function to compute a record's Maximal Matching Set (MMS) based on a single linkage structure
}

