% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeGibbs.R
\name{pairwise}
\alias{pairwise}
\title{Function to take links list that may contain 3-way, 4-way, etc.
and reduce it to pairwise only (e.g., a 3-way link 12-45-78 is
changed to 2-way links: 12-45, 12-78, 45-78}
\usage{
pairwise(.links)
}
\arguments{
\item{.links}{A vector of records that are linked to one another}
}
\value{
Returns two ways links of records
}
\description{
Function to take links list that may contain 3-way, 4-way, etc.
and reduce it to pairwise only (e.g., a 3-way link 12-45-78 is
changed to 2-way links: 12-45, 12-78, 45-78
}
