% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebGibbsSampler.R
\name{rl.gibbs}
\alias{rl.gibbs}
\title{Gibbs sampler for empirically motivated Bayesian record linkage}
\usage{
rl.gibbs(file.num = file.num, X.s = X.s, X.c = X.c, num.gs = num.gs,
  a = a, b = b, c = c, d = d, M = M)
}
\arguments{
\item{file.num}{The number of the file}

\item{X.s}{A vector of string variables}

\item{X.c}{A vector of categorical variables}

\item{num.gs}{Total number of gibb iterations}

\item{a}{Shape parameter of Beta prior}

\item{b}{Scale parameter of Beta prior}

\item{c}{Positive constant}

\item{d}{Any distance metric measuring the latent and observed string}

\item{M}{The true value of the population size}
}
\value{
Returns the estimated linkage structure via Gibbs sampling
}
\description{
Gibbs sampler for empirically motivated Bayesian record linkage
}

