% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RLdata500}
\alias{RLdata500}
\title{RLdata500}
\format{
A data frame with five variables: \code{fname_c1},\code{lname_c1},  \code{by}, code{bm}, \code{bd}.
}
\usage{
RLdata500
}
\description{
Data on synthetic generation of German names with 500 total records and 10 precent duplication.
}
\keyword{datasets}
