% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeGibbs.R
\name{check_IDs}
\alias{check_IDs}
\title{Check whether 2 records which are estimated to be linked have the same IDs}
\usage{
check_IDs(recpair, identity_vector)
}
\arguments{
\item{recpair}{A record pair}

\item{identity_vector}{A vector of the unique ids}
}
\value{
Whether or not two records which are estimated to be linked have the same unique ids
}
\description{
Check whether 2 records which are estimated to be linked have the same IDs
}
\examples{
id <- c(1,2,3,4,5,1,7,8,9,10,11,12,13,14,15,16,17,18,19,20)
rec1 <- 6
rec2 <- 1
check_IDs(recpair=c(rec1,rec2),identity_vector=id)
}
