% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebGibbsSampler.R
\name{rl.gibbs}
\alias{rl.gibbs}
\title{Gibbs sampler for empirically motivated Bayesian record linkage}
\usage{
rl.gibbs(file.num = file.num, X.s = X.s, X.c = X.c, num.gs = num.gs,
  a = a, b = b, c = c, d = d, M = M)
}
\arguments{
\item{file.num}{The number of the file}

\item{X.s}{A vector of string variables}

\item{X.c}{A vector of categorical variables}

\item{num.gs}{Total number of gibb iterations}

\item{a}{Shape parameter of Beta prior}

\item{b}{Scale parameter of Beta prior}

\item{c}{Positive constant}

\item{d}{Any distance metric measuring the latent and observed string}

\item{M}{The true value of the population size}
}
\value{
Returns the estimated linkage structure via Gibbs sampling
}
\description{
Gibbs sampler for empirically motivated Bayesian record linkage
}
\examples{
data(RLdata500)
outer_jarowinkler <- function(string1,string2) { outer(string1, string2, jarowinkler) }
d<- outer_jarowinkler
Record data for categorical fields
Note 1: The data needs to be laid out with each row representing
a record and each column representing a field.
Note 2: The different files need to be "stacked" on top of each other.
X.c <- RLdata500[c("by","bm","bd")]
X.c <- as.matrix(RLdata500[,"bd"],ncol=1)
p.c <- ncol(X.c)
X.s <- as.matrix(RLdata500[-c(2,4,7)])
p.s <- ncol(X.s)
system.time(lam.gs <- rl.gibbs(file.num,X.s,X.c,num.gs=2,a=.01,b=100,c=1,d, M=500))
rl.gibbs(file.num=rep(c(1,2,3),c(200,150,150)),X.c=as.matrix(rep(2,500)),X.s = matrix(1:12, 500,3), num.gs=2,a=.01,b=100,c=1,d=outer_jarowinkler, M=500)
}

